[Debug view] Add dynamic debug configurations
Following the recent support for dynamic debug configurations via
command "Debug: Select and Start debugging".

This change extends the functionality further to bring provided
dynamic debug configurations to the list of available configurations
under the debug view.

These configurations are refreshed when focus is gained or lost by
the "select" box, this allows to dynamically present the relevant
configurations as per:

1. A corresponding 'when' clause applicable to a specific file
extension opened in editor.

or

2) When a new provider extension has just been loaded.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>